Refactor: redux-toolkit 적용
slice.js 생성
reducer.js, actions.js 제거
store 변경
